# contracts/api/books/v1/openapi.yaml
openapi: 3.0.3
info:
  title: LoreWeave Books API
  version: 1.5.0
  description: |
    Module 02 — books, chapters (MVP text/plain .txt uploads), editable chapter draft (UTF-8) with revision history,
    cover image, summary, and per-user storage usage. Multilingual: each locale is a separate chapter row with
    original_language; list chapters supports optional query filters original_language and/or sort_order and
    pagination (limit/offset/total). Owner list includes visibility for direct badge rendering in workspace.
    Logical delete: DELETE book/chapter moves to recycle bin (trashed); DELETE …/purge from bin marks purge_pending for
    future garbage collection (blobs + DB rows removed later by a batch job, not in this API). GET /v1/books lists active
    books only; GET /v1/books/trash lists trashed books. Trashed or purge_pending books must not appear in catalog/unlisted/public.
    Raw upload bytes remain in S3-compatible storage (e.g. MinIO); draft/revisions in Postgres until purged. Gateway-facing REST only.
servers:
  - url: http://localhost:3000
    description: API Gateway (BFF)
...
paths:
...
  /v1/epub-imports:
    post:
      tags: [imports]
      summary: Start a structure-preserving EPUB import from an inspected source
      operationId: startEpubImport
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartEpubImportRequest'
      responses:
        '202':
          description: Import job queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EpubImportJobAccepted'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          description: Source/job state, duplicate source strategy, or explicit destructive confirmation is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
  /v1/import-jobs/{job_id}:
    get:
      tags: [imports]
      summary: Get EPUB import job status, progress, and allowed recovery actions
      operationId: getEpubImportJob
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ImportJobId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EpubImportJob'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
...
  /v1/books/{book_id}/import:
    post:
      tags: [chapters]
      summary: Import a whole book file (.docx, .epub, .fb2, .md, .pdf, .txt) as chapters into an existing book
      description: >
        docs/specs/2026-07-06-pdf-book-import.md. .txt is parsed synchronously (source_format=plain)
        and returns the created chapter(s) directly. All other formats are stored and processed
        asynchronously by worker-infra — poll GET .../imports/{import_id} or listen on the
        `/ws` import-progress channel. For file_format=pdf with caption_images=true, model_source
        and model_ref are REQUIRED (Provider gateway invariant — no platform default vision model;
        the caller must explicitly pick a vision-capable BYOK model).
      operationId: startImport
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BookId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: .txt, .docx, .epub, .fb2, .md, or .pdf — up to 200 MB.
                original_language:
                  type: string
                  maxLength: 35
                  description: BCP-47 tag (e.g. vi, zh-Hans). Defaults to "auto" when omitted.
                pages_per_chunk:
                  type: integer
                  minimum: 1
                  description: Required (and only read) when the file is a PDF — how many source pages become one chapter.
                caption_images:
                  type: boolean
                  description: PDF only. When true, model_source/model_ref become required.
                model_source:
                  type: string
                  description: PDF + caption_images only. BYOK provider source for the vision model (e.g. "user_model").
                model_ref:
                  type: string
                  description: PDF + caption_images only. The user_model_id (UUID) of the vision-capable model to caption images with.
      responses:
        '201':
          description: >
            .txt path only — synchronous. A single Chapter when the file has one part,
            or a BulkImportSummary when it decomposed into multiple parts/chapters.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Chapter'
                  - $ref: '#/components/schemas/BulkImportSummary'
        '202':
          description: .docx/.epub/.fb2/.md/.pdf — queued for async processing by worker-infra. FB2 metadata is retained with the import but does not overwrite existing book metadata.
...
  /v1/books/{book_id}/imports:
    get:
      tags: [chapters]
      summary: List recent import jobs for this book (most recent 20)
      operationId: listImportJobs
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BookId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportJobListResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
  /v1/books/{book_id}/imports/{import_id}:
    get:
      tags: [chapters]
      summary: Get one import job's status/progress
      operationId: getImportJob
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BookId'
        - name: import_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportJob'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    BookId:
      name: book_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    ImportJobId:
      name: job_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    ErrorResponse:
      description: Error envelope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
  schemas:
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
    EpubImportWarning:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - navigation_missing_spine_fallback
            - navigation_anchor_missing
            - navigation_range_ambiguous
            - navigation_node_non_content
            - unsupported_resource
            - asset_missing
            - asset_invalid_mime
            - internal_link_target_missing
            - metadata_conflict
            - composition_materialization_failed
            - rollback_conflict_user_modified
        message:
          type: string
        source_key:
          type: string
          nullable: true
    EpubImportErrorCode:
      type: string
      enum:
        - import_source_missing
        - epub_invalid_archive
        - epub_mimetype_invalid
        - epub_container_missing
        - epub_opf_invalid
        - epub_drm_unsupported
        - epub_archive_path_invalid
        - epub_archive_limit_exceeded
        - epub_navigation_invalid
        - epub_item_processing_failed
        - import_cancelled
        - import_state_conflict
...
    EpubImportJobStatus:
      type: string
      enum:
        - queued
        - processing
        - import_staging
        - import_ready
        - completed
        - completed_with_warnings
        - import_failed
        - cancelled
        - rolling_back
        - rolled_back
...
    EpubImportTarget:
      type: object
      required: [mode]
      properties:
        mode:
          type: string
          enum: [existing_book, new_book]
        book_id:
          type: string
          format: uuid
          nullable: true
          description: Required when mode is existing_book and omitted when mode is new_book.
    EpubMetadataPolicy:
      type: object
      properties:
        title:
          type: string
          enum: [use_source, keep_existing, edit_manually]
        authors:
          type: string
          enum: [use_source, keep_existing, merge, edit_manually]
        language:
          type: string
          enum: [use_source, keep_existing, edit_manually]
        description:
          type: string
          enum: [use_source, keep_existing, merge, edit_manually]
        cover:
          type: string
          enum: [use_source, keep_existing]
    EpubImportOptions:
      type: object
      properties:
        preserve_hierarchy:
          type: boolean
          default: true
        import_images:
          type: boolean
          default: true
        import_footnotes:
          type: boolean
          default: true
        retain_source:
          type: boolean
          default: true
    StartEpubImportRequest:
      type: object
      required: [source_id, target, strategy, metadata_policy, options]
      properties:
        source_id:
          type: string
          format: uuid
        target:
          $ref: '#/components/schemas/EpubImportTarget'
        strategy:
          type: string
          enum: [append, replace_all, merge_by_source_key]
        destructive_confirmation:
          type: boolean
          description: Required and true when strategy is replace_all.
        metadata_policy:
          $ref: '#/components/schemas/EpubMetadataPolicy'
        selected_source_keys:
          type: array
          description: Empty means all selectable leaf items from the inspection.
          items:
            type: string
        title_overrides:
          type: object
          additionalProperties:
            type: string
        options:
          $ref: '#/components/schemas/EpubImportOptions'
    EpubImportJobAccepted:
      type: object
      required: [job_id, source_id, status]
      properties:
        job_id:
          type: string
          format: uuid
        source_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued]
    EpubImportProgress:
      type: object
      required: [total, completed, failed]
      properties:
        total:
          type: integer
          minimum: 0
        completed:
          type: integer
          minimum: 0
        failed:
          type: integer
          minimum: 0
        current_source_key:
          type: string
          nullable: true
        assets_imported:
          type: integer
          minimum: 0
    EpubImportJob:
      type: object
      required: [job_id, source_id, status, progress, warnings, created_at, updated_at, resumable, cancellable, rollback_available]
      properties:
        job_id:
          type: string
          format: uuid
        source_id:
          type: string
          format: uuid
        book_id:
          type: string
          format: uuid
          nullable: true
        status:
          $ref: '#/components/schemas/EpubImportJobStatus'
        progress:
          $ref: '#/components/schemas/EpubImportProgress'
        warnings:
          type: array
          items:
            $ref: '#/components/schemas/EpubImportWarning'
        error_code:
          allOf:
            - $ref: '#/components/schemas/EpubImportErrorCode'
          nullable: true
        error_message:
          type: string
          nullable: true
        resumable:
          type: boolean
        cancellable:
          type: boolean
        rollback_available:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        finalized_at:
          type: string
          format: date-time
          nullable: true
...
    LifecycleState:
      type: string
      enum: [active, trashed, purge_pending]
      description: >
        active — normal workspace; trashed — recycle bin (logical delete stage 1);
        purge_pending — user removed from bin (stage 2); data kept until garbage collector deletes storage and rows.
...
    Chapter:
      type: object
      required:
        - chapter_id
        - book_id
        - original_filename
        - original_language
        - content_type
        - byte_size
        - sort_order
        - lifecycle_state
        - created_at
        - updated_at
      properties:
        chapter_id:
          type: string
          format: uuid
        book_id:
          type: string
          format: uuid
        title:
          type: string
        original_filename:
          type: string
        original_language:
          type: string
          maxLength: 35
          description: BCP-47 for this chapter's uploaded/edited content; user may upload another chapter row per language.
        content_type:
          type: string
          example: text/plain
        byte_size:
          type: integer
          minimum: 0
          description: Size of original uploaded object in storage (not draft text length).
        sort_order:
          type: integer
        draft_updated_at:
          type: string
          format: date-time
          description: When canonical draft text was last saved (if applicable).
        draft_revision_count:
          type: integer
          minimum: 0
          description: Number of stored revision snapshots for this chapter.
        lifecycle_state:
          $ref: '#/components/schemas/LifecycleState'
        trashed_at:
          type: string
          format: date-time
          nullable: true
        purge_eligible_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    BulkImportSummary:
      type: object
      required: [chapters_created, book_id]
      description: Returned by POST .../import (text path) when the file decomposed into more than one chapter.
      properties:
        chapters_created:
          type: integer
          minimum: 0
        book_id:
          type: string
          format: uuid
...
    ImportJob:
      type: object
      required: [id, book_id, status, filename, file_format, file_size, chapters_created, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        book_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, processing, completed, failed]
        filename:
          type: string
        file_format:
          type: string
          enum: [docx, epub, fb2, markdown, pdf]
        file_size:
          type: integer
          minimum: 0
        chapters_created:
          type: integer
          minimum: 0
        error:
          type: string
          nullable: true
        created_at:
          type: string
        updated_at:
          type: string
        completed_at:
          type: string
          nullable: true
    ImportJobListResponse:
      type: object
      required: [imports]
      properties:
        imports:
          type: array
          items:
            $ref: '#/components/schemas/ImportJob'
...